Moved message event to own file, attached list and config to client
